\name{StatsProgramming-package}
\alias{StatsProgramming-package}
\alias{StatsProgramming}
\docType{package}
\title{
\packageTitle{StatsProgramming}
}
\description{
\packageDescription{StatsProgramming}
}
\details{
The package has the following four functions: 

1) myQR: Performs QR decomposition on the matrix A.

2) myLM: Performs the linear regression of Y on X. 

3) myEigen_QR: Performs PCA on matrix A using  QR Decomposition .

4) myLogistic: Performs the logistic regression of Y on X.
}
\author{
\packageAuthor{StatsProgramming}

Maintainer: \packageMaintainer{StatsProgramming}
}

\seealso{

 \code{\link{myQR}},
 \code{\link{myLM}},
 \code{\link{myEigen_QR}},
 \code{\link{myLogistic}},
}
\examples{
}
\references{
	\href{https://en.wikipedia.org/wiki/QR_decomposition}{QR decomposition},
	\href{https://en.wikipedia.org/wiki/Linear_regression}{Linear regression},
	\href{https://en.wikipedia.org/wiki/Principal_component_analysis}{Principal component analysis},
	\href{https://en.wikipedia.org/wiki/Logistic_regression}{Logistic regression}
}